Remove `PBuiltinMap`
This type alias was confusing to users, since it was not clear what the difference
between this and `PMap` was. `PMap` was a newtype of `PBuiltinMap`, but almost in any
case where you have `PBuiltinMap`, you really want the newtype `PBuiltinMap`.

The issue is that there are instances on `PBuiltinMap` that work very differently
from the ones on `PBuiltinList`, even though it's just a type alias!
When these instances are instead made to be instances on the newtype,
then there's not much point in keeping `PBuiltinMap` around, especially
when it only serves to increase confusion.

Closes #513